+++ src/api/lib/domain/translator/index.ts
@@ -1,0 +1,5 @@
+export * from "./Language";
+export * from "./OriginalText";
+export * from "./TranslatedText";
+export * from "./TranslationConfig";
+export * from "./TranslationResult";
