+++ src/api/lib/domain/user/index.ts
@@ -1,0 +1,3 @@
+export * from "./auth";
+export * from "./User";
+export * from "./UserId";
