+++ src/api/usecase/auth/SignUpUsecase.ts
@@ -1,4 +1,4 @@
-import { UserAuth } from "@/api/lib/domain/user/auth/UserAuth";
+import { UserAuth } from "@/api/lib/domain";
 import { userService } from "@/api/service";
 
 export const SignUpUsecase = async (auth: UserAuth) => {
